requirements.txt
# Monday.com BI Agent Requirements
# Install with: pip install -r requirements.txt

# Core Dependencies
openai>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0

# Security & Environment
python-dotenv>=1.0.0
